refactor: Split up widgets to files
