+++ packages/ubuntu_bootstrap/lib/slides/widgets/slide_link.dart
@@ -1,0 +1,26 @@
+import 'package:flutter_html/flutter_html.dart';
+import 'package:flutter/material.dart';
+import 'package:ubuntu_utils/ubuntu_utils.dart';
+
+class SlideLink extends StatelessWidget {
+  const SlideLink({required this.text, required this.url});
+
+  final String text;
+  final String url;
+
+  @override
+  Widget build(BuildContext context) {
+    return Html(
+      data: '<a href="$url">$text</a>',
+      shrinkWrap: true,
+      style: {
+        'body': Style(margin: Margins.zero),
+        'a': Style(
+          color: Theme.of(context).colorScheme.onBackground,
+          textDecoration: TextDecoration.none,
+        ),
+      },
+      onAnchorTap: (url, _, __) => launchUrl(url!),
+    );
+  }
+}
